feat: first attempt at compiler agnosticism
These changes allow the project to be compiled with a non MS compiler.

MSVC seems to not enforce the C++17 standard correctly and allows dynamic exceptions to be specified, so we use C++14 instead.

Additionally, there is a timezone struct that is defined within Micro-Manager when using MSVC.

Currently, this doesn't work yet as some symbols seem to be passed incorrectly.
When I compile with this branch, Micro-Manager lets you choose the device in the hardware wizard,
but gets the wrong type enum.